% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/actors.R
\name{actors}
\alias{actors}
\alias{subtitle}
\title{Collect info about actors}
\usage{
actors(www)
}
\arguments{
\item{www}{A link - see full cast from imdb.}
}
\value{
list with info about actors, moreover all info is saved in csv file
}
\description{
Function \code{actors} gets info about actors (age,place and date of birth, death).
}

