% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aktorzy_wyszukaj_wiek_kraj.R
\name{aktorzy_wyszukaj_wiek_kraj}
\alias{aktorzy_wyszukaj_wiek_kraj}
\title{Funkcja dla danego filmu, wyszukuje pierwszych 10 aktorow w ramce danych aktorzy.}
\usage{
aktorzy_wyszukaj_wiek_kraj(f, aktorzy)
}
\arguments{
\item{f}{ramka danych z informacja o jednym filmie}

\item{aktorzy}{ramka danych z aktorami}
}
\value{
lista z informacjami o aktorach, pierwszy element listy to ich wiek w momencie
grania w filmie, natomiast drugi element to wektor krajow pochodzenia
}
\description{
Funkcja \code{aktorzy_wyszukaj_wiek_kraj} dla danego filmu, wyszukuje pierwszych
10 aktorow w ramce danych aktorzy (wyciaga takie informacje jak ich wiek w momencie
grania w filmie i kraj pochodzenia)
}
\author{
Emilia Momotko
}

