% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\alias{subtitle}
\title{Create a file from a data.frame and saves it.}
\usage{
create(fname, lista)
}
\arguments{
\item{fname}{Name of a file where info is to be saved.}

\item{frame}{a data frame with 5 colums}
}
\value{
invisible(NULL)  - all info is gathered in a particular file.
}
\description{
Function \code{create} saves info about actors in a file.
}

