% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create3.R
\name{create2}
\alias{create2}
\alias{subtitle}
\title{Create a file from a data.frame and saves it.}
\usage{
create2(fname, lista)
}
\arguments{
\item{fname}{Name of a file where info is to be saved.}

\item{frame}{a data frame with 6 colums}
}
\value{
invisible(NULL)  - all info is gathered in a particular file.
}
\description{
Function \code{create} gets info about actors.
}

