% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/czas.R
\name{czas}
\alias{czas}
\title{Funkcja sprawdza czy dwie wartosci liczbowe sa "podobne"}
\usage{
czas(czas_glowny, czas_por)
}
\arguments{
\item{czas_glowny}{pierwsza wartosc numeryczna, liczba naturalna}

\item{czas_por}{druga wartosc numeryczna, liczba naturalna}
}
\value{
wartosc numeryczna 1, jesli istnieje istotne podobienstwo miedzy wartosciami, 0 wpp
}
\description{
Funkcja \code{czas} sprawdza czy dwie wartosci liczbowe sa "podobne" (w domysle
chodzi o czas trwania filmu)
}
\author{
Emilia Momotko
}

