% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/czy_to_samo.R
\name{czy_to_samo}
\alias{czy_to_samo}
\title{Funkcja okreslajaca czy dwa filmy otrzymały te same nagrody, czy nie.}
\usage{
czy_to_samo(text_gl, text_por)
}
\arguments{
\item{text_gl}{liczba naturalna okreslajaca liczbe nagrod filmu glownego}

\item{text_por}{liczba naturalna okreslajaca liczbe nagrod filmu porownawczego}
}
\value{
wartosc numeryczna 0 lub 1.
}
\description{
Funkcja \code{czy_to_samo} okresla czy dwa filmy porownanwcze otrzymaly
te same nagrody np. Oscary
}
\author{
Adrianna Sudol
}

