% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/czy_zawiera.R
\name{czy_zawiera}
\alias{czy_zawiera}
\title{Funkcja sprawdzajaca czy jakies wartosci pewnej cechy filmu powtarzaja sie.}
\usage{
czy_zawiera(glowny, por)
}
\arguments{
\item{glowny}{napis okreslajacy ceche filmu glownego, ktorego separatorem jest}

\item{por}{napis okreslajacy ceche filmu porownawczego, ktorego separatorem jest}
}
\value{
wartosc numeryczna 0 lub 1.
}
\description{
Funkcja \code{czy_zawiera} sprawdza czy elementej jednej cechy filmu zawiera sie
w tej samej cesze drugiego flimu np. czy kraje powstania filmu sie powtarzaja.
}
\author{
Adrianna Sudol
}

