% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/director.R
\name{director}
\alias{director}
\alias{subtitle}
\title{Collect info about directors}
\usage{
director(www)
}
\arguments{
\item{www}{A link - see full cast from imdb.}
}
\value{
list with info about actors, moreover all info is saved in csv file.
}
\description{
Function \code{director} gets info about directors.
}

