% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dodaj_aktorzy_rezyser.R
\name{dodaj_aktorzy_rezyser}
\alias{dodaj_aktorzy_rezyser}
\title{Funkcja do ramki danych filmy dodaje kolumny zwiazanie z odpowiednimi wskaznikami
zwiazanymi z obsada i rezyseria}
\usage{
dodaj_aktorzy_rezyser(filmy, aktorzy, rezyserzy)
}
\arguments{
\item{filmy}{ramka danych z filmami}

\item{aktorzy}{ramka danych z aktorami}

\item{rezyserzy}{ramka danych z rezyserami}
}
\value{
ramka danych z dodanymi kolumnami
}
\description{
Funkcja \code{dodaj_aktorzy_rezyser} do ramki danych filmy dodaje kolumny zwiazanie z odpowiednimi wskaznikami
zwiazanymi z obsada i rezyseria (rozrzut wieku, mediane wieku, kraje pochodzenia)
}
\author{
Emilia Momotko
}

