% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frakcja.R
\name{frakcja}
\alias{frakcja}
\title{Funkcja wyznacza wskaznik podobienstwa miedzy dwoma wartosciami tekstowymi}
\usage{
frakcja(a1, a2)
}
\arguments{
\item{a1}{tekst pierwszy, typu character (wyrazenia rozdzielone 'malpa')}

\item{a2}{tekst drugi, typu character (wyrazenia rozdzielone 'malpa')}
}
\value{
wartosc numeryczna z przedzialu [0,1] - podobiestwo obu tekstow
}
\description{
Funkcja \code{frakcja} wskaznik podobienstwa miedzy dwoma wartosciami tekstowymi
przy zalozeniu ze slowa lub wyrazenia z jednego tekstu sa odseparowane 'malpa'
}
\author{
Emilia Momotko
}

