% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frakcja_powtarzanych_slow.R
\name{frakcja_powtarzanych_slow}
\alias{frakcja_powtarzanych_slow}
\title{Funkcja wyznacza odpowiednie podobienstwo miedzy dwoma dowolnymi tekstami}
\usage{
frakcja_powtarzanych_slow(text_gl, text_por)
}
\arguments{
\item{text_gl}{wartosc tekstowa, napis}

\item{text_por}{wartosc tekstowa, napis}
}
\value{
wartosc numeryczna z przedzialu [0,1] - podobienstwo miedzy dwoma tekstami
}
\description{
Funkcja \code{frakcja_powtarzanych_slow} wyznacza odpowiednie podobienstwo
miedzy dwoma dowolnymi tekstami
}
\author{
Emilia Momotko
}

