% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\alias{subtitle}
\title{Collect names of producers and musicians}
\usage{
get_names(www)
}
\arguments{
\item{www}{A link - see full cast from imdb.}
}
\value{
a list: first element - musicians, second - producers
}
\description{
Function \code{get_names} gets names of director and producers.
}

