% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/harvest_people.R
\name{harvest_people}
\alias{harvest_people}
\alias{subtitle}
\title{Collect info about people}
\usage{
harvest_people(htmls)
}
\arguments{
\item{htmls}{list of html objects - parsed HTML pages .}
}
\value{
list with info about people.
}
\description{
Function \code{harvest_people} gets info about people.
}

