% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/htmls_movie.R
\name{htmls_movie}
\alias{htmls_movie}
\alias{subtitle}
\title{Parse many of HTML pages for actorS AND director}
\usage{
htmls_movie(www, selector)
}
\arguments{
\item{www}{A link - see full cast from imdb.}

\item{selector}{named character vector with selectors which we are going to analyse.}
}
\value{
list of parsed pages or 'NA' if pages don't exist.
}
\description{
Function \code{htmls_movie} parses html pages
}

