% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeListOfUrls.R
\name{makeListOfUrls}
\alias{makeListOfUrls}
\title{Funkcja pobiera linki do filmów z bazy IMdB}
\usage{
makeListOfUrls(urlStart)
}
\arguments{
\item{urlStart,}{początkowy url}
}
\description{
Funkcja pobiera linki do filmów z bazy IMdB
}
\examples{
# ściągamy filmy z kilku list - MovieMeter, Numer of Votes, US Box Office
# dzięki temu będziemy mieć znacznie więcej niż 100 000, bo filmy są w różnej kolejności
# więc parsujemy trzy podstrony i bierzemy niepowtarzające się filmy

tab <- makeListOfUrls('http://www.imdb.com/search/title?at=0&sort=moviemeter,asc&start=1&title_type=feature')
write.table(tab, 'linki_tmp.txt', row.names=FALSE, col.names=FALSE)

tab2 <- makeListOfUrls('http://www.imdb.com/search/title?at=0&sort=num_votes&start=1&title_type=feature')
write.table(tab6, 'linki_tmp2.txt', row.names=FALSE, col.names=FALSE)

tab3 <- makeListOfUrls('http://www.imdb.com/search/title?at=0&sort=boxoffice_gross_us&start=1&title_type=feature')
write.table(tab3, 'linki_tmp3.txt', row.names=FALSE, col.names=FALSE)
}

