% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odleglosc.R
\name{odleglosc}
\alias{odleglosc}
\title{Funkcja zlicza odleglosc (podobienstwo) miedzy jednym a drugim filmem}
\usage{
odleglosc(f_glowny, f_por, filmy)
}
\arguments{
\item{f_glowny}{- film glowny, czyli jednowierszowa ramka danych zawierajaca odpowiednie info}

\item{f_por}{- film porownywany, czyli jednowierszowa ramka danych zawierajaca odpowiednie info}

\item{filmy}{- ramka danych z filmami}
}
\value{
wartosc numeryczna z przedzialu [0,1] - podobienstwo miedzy filmami
}
\description{
Funkcja \code{odleglosc} zlicza wartosci podobienstw meidzy dwoma zadanymi filmami.
Podobienstwo to jest wyznaczane na podstawie wartosci wskaznikow: roku powstania,
gatunkow, kraju produkcji, czasu trwania, autorow muzyki, producentow, aktorow, rezysera,
ocen, liczby oceniajacych uzytkownikow, liczby recenzji, tresci recenzji, slow kluczowych,
tresci opisu, tresci fabuly, liczby oscarow, liczby innych nagrod, mediany wieku aktorow,
zroznicowania wieku aktorow, wieku rezysera, kraju pochodzenia rezysera i aktorow.
}
\author{
Emilia Momotko
}

