% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/porownaj_mediany.R
\name{porownaj_mediany}
\alias{porownaj_mediany}
\title{Funkcja sprawdza czy dwie wartosci sa odpowiednio sobie bliskie}
\usage{
porownaj_mediany(m1, m2)
}
\arguments{
\item{m1}{liczba naturalna - mediana pierwszej grupy badanych osob}

\item{m2}{liczba naturalna - mediana drugiej grupy badanych osob}
}
\value{
wartosc numeryczna 1, jesli mediany sa zblizone, 0 wpp.
}
\description{
Funkcja \code{porownaj_mediany} sprawdza czy dwie wartosci liczbowe (w domysle sa to
mediany wieku) nie roznia sie zbytnio.
}
\author{
Emilia Momotko
}

