% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/porownaj_rozrzut.R
\name{porownaj_rozrzut}
\alias{porownaj_rozrzut}
\title{Funkcja wyznacza wskaznik podobienstwa miedzy dwiema wartosciami rozrzutu}
\usage{
porownaj_rozrzut(r1, r2)
}
\arguments{
\item{r1}{liczba naturalna - rozrzut pierwszej grupy badanych osob}

\item{r2}{liczba naturalna - rozrzut drugiej grupy badanych osob}
}
\value{
wartosc numeryczna z przedzialu [0,1] - podobiestwo wartosci rozrzutu w grupie wiekowej
}
\description{
Funkcja \code{porownaj_rozrzut} wyznacza wskaznik podobienstwa miedzy dwiema
wartosciami rozrzutu (w domysle chodzi o rozrzut wieku badanej grupy osob)
}
\author{
Emilia Momotko
}

