% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rezyser_wyszukaj_wiek_kraj.R
\name{rezyser_wyszukaj_wiek_kraj}
\alias{rezyser_wyszukaj_wiek_kraj}
\title{Funkcja dla danego filmu, wyszukuje pierwszych rezyserow w ramce danych rezyserzy}
\usage{
rezyser_wyszukaj_wiek_kraj(f, rezyserzy)
}
\arguments{
\item{f}{ramka danych z informacja o jednym filmie}

\item{rezyserzy}{ramka danych z rezyserami}
}
\value{
lista z informacjami o rezyserze lub rezyserach: pierwszy element listy to ich wiek w momencie
grania w filmie, natomiast drugi element to wektor krajow pochodzenia
}
\description{
Funkcja \code{rezyser_wyszukaj_wiek_kraj} dla danego filmu, wyszukuje rezyserow w ramce
danych rezyserzy (wyciaga takie informacje jak ich wiek w momencie
rezyserowania filmu i kraj pochodzenia)
}
\author{
Emilia Momotko
}

