% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rok.R
\name{rok}
\alias{rok}
\title{Funkcja okreslajaca czy lata powstania fillmu sa zblizone.}
\usage{
rok(rok_glowny, rok_por)
}
\arguments{
\item{rok_glowny}{liczba naturalna (jako wektor napisow) okreslajaca rok wydania filmu glownego}

\item{por}{liczba naturalna (jako wektor napisow) okreslajaca rok wydania filmu porownawczego}
}
\value{
wartosc numeryczna 0 lub 1.
}
\description{
Funkcja \code{rok} okresla czy dwa porwnawcze filmy sa podobne pod wzgledem lat
powstania. Uwzgledniony jest fakt ze roznica miedzy rokiem 1942 a 1948 jest inna niz
miedzy 2005 a 2011, gdyz wiadomo postep techniki byl inny.
}
\author{
Adrianna Sudol
}

