% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roznica.R
\name{roznica}
\alias{roznica}
\title{Funkcja okreslajaca cechy filmow zawierają się w oczekiwanym przedziale wartosci}
\usage{
roznica(glowny, por,param)
}
\arguments{
\item{glowny}{liczba naturalna (jako wektor napisow) okreslajaca ceche filmu glownego}

\item{por}{liczba naturalna (jako wektor napisow) okreslajaca ceche filmu porownawczego}

\item{param}{wartosc liczbowa okreslajaca max jakim moga sie roznic dane cechy, dla
       oceny bedzie 0.3, liczby uzytkownikow 50000, liczby recenzji 20}
}
\value{
wartosc numeryczna 0 lub 1.
}
\description{
Funkcja \code{roznica} okresla czy pewna cecha dla dwoch filmow jest podobna pod
wzgledem zawierania sie w podobnych przedzialach. Sluzy do porownania ocen filmu,
liczby uzytkownikow oceniajacych oraz liczby recenzji.
}
\author{
Adrianna Sudol
}

