% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sprawdz.R
\name{sprawdz}
\alias{sprawdz}
\title{Checks whether the value is NA and converts it to a string.}
\usage{
sprawdz(cecha)
}
\arguments{
\item{cecha}{Vector}
}
\value{
Character vector
}
\description{
Function \code{wyjmij} checks whether the value is NA and converts it to a string.
}

