% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Collect info about summary}
\usage{
summary(links, titles, film)
}
\arguments{
\item{links}{A link - see full summary from imdb.}

\item{titles}{Titles references from imdb.}

\item{film}{Either a complete document (XMLInternalDocument), a list of tags (XMLNodeSet) or a single tag (XMLInternalElementNode)}
}
\value{
Character vector
}
\description{
Function \code{summary} gets info about actors.
}

