% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tworz_macierz.R
\name{tworz_macierz}
\alias{tworz_macierz}
\title{Funkcja tworzy macierz podobienstw}
\usage{
tworz_macierz(ile, filmy)
}
\arguments{
\item{ile}{liczba naturalna, dodatnia, liczba porownywanych filmow}

\item{filmy}{ramka danych z filmami}
}
\value{
macierz rozmiaru ile x ile
}
\description{
Funkcja \code{tworz_macierz} tworzy macierz podobienstw na podstawie zadanej
liczby filmow do przeanalizowania i ramki danych z filmami
}
\author{
Emilia Momotko
}

