% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wszystko.R
\name{wszystko}
\alias{wszystko}
\title{Collect all info about film,}
\usage{
wszystko(link)
}
\arguments{
\item{link}{A link from imdb.}
}
\value{
List
}
\description{
Function \code{wszystko} gets all info about film.
}

