% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wydobadz.R
\name{wydobadz}
\alias{wydobadz}
\title{Collect any information.}
\usage{
wydobadz(links, titles, film, nodes1, nodes2)
}
\arguments{
\item{links}{A link - see full summary from imdb.}

\item{titles}{Titles references from imdb.}

\item{film}{Either a complete document (XMLInternalDocument), a list of tags (XMLNodeSet) or a single tag (XMLInternalElementNode)}

\item{nodes1}{Nodes to select}

\item{nodes2}{Nodes to select}
}
\value{
Character vector
}
\description{
Function \code{wydobadz} gets info about something .
}

