% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wyjmij.R
\name{wyjmij}
\alias{wyjmij}
\title{Collect any information.}
\usage{
wyjmij(film, nodes)
}
\arguments{
\item{film}{Either a complete document (XMLInternalDocument), a list of tags (XMLNodeSet) or a single tag (XMLInternalElementNode)}

\item{nodes}{Nodes to select}
}
\value{
Character vector
}
\description{
Function \code{wyjmij} gets info about something .
}

