% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create2.R
\name{create2}
\alias{create2}
\alias{subtitle}
\title{Create a file from a data.frame and saves it.}
\usage{
create2(fname, lista)
}
\arguments{
\item{fname}{Name of a file where info is to be saved.}

\item{lista}{a data frame with 24 colums}
}
\value{
invisible(NULL)  - all info is gathered in a particular file.
}
\description{
Function \code{create} gets all informations about film.
}

